refactor project struct
