+++ app/src/main/java/org/ciaranroche/kang/fragments/main/VinylProfileFragment.kt
@@ -1,4 +1,4 @@
-package org.ciaranroche.kang.fragments
+package org.ciaranroche.kang.fragments.main
 
 import android.net.Uri
 import android.os.Bundle
