+++ app/src/main/res/layout/fragment_add_vinyl.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.ciaranroche.kang.fragments.AddVinylFragment">
+    tools:context="org.ciaranroche.kang.fragments.main.AddVinylFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
