+++ app/src/main/res/layout/fragment_log_in.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".fragments.LogInFragment"
+                                                   tools:context=".fragments.startup.LogInFragment"
                                                    android:id="@+id/frameLayout5"
                                                    android:background="@color/colorPrimaryDark">
