+++ app/src/main/res/layout/fragment_log_out.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".fragments.LogOutFragment"
+                                                   tools:context=".fragments.main.LogOutFragment"
                                                    android:id="@+id/frameLayout4"
                                                    android:background="@color/colorPrimaryDark">
