+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".fragments.SignUpFragment"
+                                                   tools:context=".fragments.startup.SignUpFragment"
                                                    android:id="@+id/frameLayout6"
                                                    android:background="@color/colorPrimaryDark">
