+++ app/src/main/res/layout/fragment_thank_you.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragments.ThankYouFragment"
+        tools:context=".fragments.main.ThankYouFragment"
         android:id="@+id/frameLayout2"
         android:background="@color/colorPrimaryDark">
