+++ app/src/main/res/layout/fragment_vinyl_list.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context="org.ciaranroche.kang.fragments.HomeFragment">
+                                                   tools:context="org.ciaranroche.kang.fragments.main.HomeFragment">
 
 
     <com.nshmura.recyclertablayout.RecyclerTabLayout
