+++ app/src/main/res/layout/fragment_vinyl_more.xml
@@ -5,13 +5,13 @@
             xmlns:tools="http://schemas.android.com/tools"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".fragments.VinylMoreFragment">
+            tools:context=".fragments.main.VinylMoreFragment">
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
                                                    xmlns:app="http://schemas.android.com/apk/res-auto"
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".fragments.VinylMoreFragment"
+                                                   tools:context=".fragments.main.VinylMoreFragment"
                                                    android:id="@+id/frameLayout3">
 
     <TextView
