+++ app/src/main/res/layout/fragment_welcome.xml
@@ -5,7 +5,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".fragments.WelcomeFragment"
+tools:context=".fragments.main.WelcomeFragment"
 android:id="@+id/frameLayout">
 
     <androidx.recyclerview.widget.RecyclerView
