+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,7 +6,7 @@
         >
     <fragment
             android:id="@+id/welcomeFragment"
-            android:name="org.ciaranroche.kang.fragments.WelcomeFragment"
+            android:name="org.ciaranroche.kang.fragments.main.WelcomeFragment"
             android:label="Select Kang Genre"
             tools:layout="@layout/fragment_welcome">
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
             android:id="@+id/vinylProfileFragment"
-            android:name="org.ciaranroche.kang.fragments.HomeFragment"
+            android:name="org.ciaranroche.kang.fragments.main.HomeFragment"
             android:label="Kang"
             tools:layout="@layout/fragment_vinyl_list">
         <action
@@ -47,7 +47,7 @@
 
     <fragment
             android:id="@+id/add_vinyl"
-            android:name="org.ciaranroche.kang.fragments.AddVinylFragment"
+            android:name="org.ciaranroche.kang.fragments.main.AddVinylFragment"
             android:label="@string/add_vinyl"
             tools:layout="@layout/fragment_add_vinyl">
         <argument
@@ -66,14 +66,14 @@
 
     <fragment
             android:id="@+id/settings_dest"
-            android:name="org.ciaranroche.kang.fragments.SettingsFragment"
+            android:name="org.ciaranroche.kang.fragments.main.SettingsFragment"
             android:label="@string/settings"
             tools:layout="@layout/fragment_settings" />
 
 
     <fragment
             android:id="@+id/deeplink_dest"
-            android:name="org.ciaranroche.kang.fragments.DeepLinkFragment"
+            android:name="org.ciaranroche.kang.fragments.main.DeepLinkFragment"
             android:label="@string/deeplink"
             tools:layout="@layout/deeplink_fragment">
 
@@ -85,7 +85,7 @@
     </fragment>
     <fragment
             android:id="@+id/thankYouFragment"
-            android:name="org.ciaranroche.kang.fragments.ThankYouFragment"
+            android:name="org.ciaranroche.kang.fragments.main.ThankYouFragment"
             android:label="Thank You"
             tools:layout="@layout/fragment_thank_you">
         <action
@@ -98,7 +98,7 @@
     </fragment>
     <fragment
             android:id="@+id/vinylMoreFragment"
-            android:name="org.ciaranroche.kang.fragments.VinylMoreFragment"
+            android:name="org.ciaranroche.kang.fragments.main.VinylMoreFragment"
             android:label="About Vinyl"
             tools:layout="@layout/fragment_vinyl_more">
         <action
@@ -111,7 +111,7 @@
     </fragment>
     <fragment
             android:id="@+id/noVinylFragment"
-            android:name="org.ciaranroche.kang.fragments.NoVinylFragment"
+            android:name="org.ciaranroche.kang.fragments.main.NoVinylFragment"
             android:label="No Vinyl"
             tools:layout="@layout/fragment_no_vinyl">
         <action
@@ -124,7 +124,7 @@
     </fragment>
     <fragment
             android:id="@+id/logOutFragment"
-            android:name="org.ciaranroche.kang.fragments.LogOutFragment"
+            android:name="org.ciaranroche.kang.fragments.main.LogOutFragment"
             android:label="Log Out"
             tools:layout="@layout/fragment_log_out">
         <action
